+++ src/include/codegen/util/index_scan_iterator.h
@@ -42,8 +42,8 @@ class IndexScanIterator {
   uint64_t GetDistinctTileGroupNum() {
     return (uint64_t)distinct_tile_group_num_;
   }
-  uint64_t GetTileGroupId(uint64_t distinct_tile_index) {
-    return (uint64_t)(result_[distinct_tile_index]->block);
+  uint32_t GetTileGroupId(uint64_t distinct_tile_index) {
+    return (uint32_t)(result_[distinct_tile_index]->block);
   }
   uint32_t GetTileGroupOffset(uint64_t result_iter) {
     return (uint32_t)(result_[result_iter]->offset);
